requirements.txt
matplotlib==3.5.3
numpy==1.23.2
pandas<1.4.4
Pillow==9.2.0
scikit_learn==1.1.2
seaborn==0.12.0
streamlit==1.12.2
